requirements.txt
Flask
together
python-dotenv
requests
Pillow
cloudinary
pymongo
torch
transformers
torchvision
pandas
numpy
datasets 
accelerate
dnspython